removed gratuitous motor sign type conversion

+++ src/motor.rs
@@ -18,7 +18,7 @@ where
         let sign = if speed >= 0 { 1 } else { 2 };
         // XXX Fixme: Return error on overspeed.
         let speed = speed.abs().min(100);
-        let buf = [motor, sign as u8, speed as u8, 0];
+        let buf = [motor, sign, speed as u8, 0];
         self.i2c.write(Self::I2C_ADDR, &buf)
     }
 }
